move cancelOrder inventory adjustments into appEvent
