+++ imports/plugins/core/orders/server/methods/cancelOrder.js
@@ -11,6 +11,8 @@ import updateCatalogProductInventoryStatus from "/imports/plugins/core/catalog/s
 import updateParentVariantsInventoryAvailableToSellQuantity from "/imports/plugins/core/inventory/server/no-meteor/utils/updateParentVariantsInventoryAvailableToSellQuantity";
 import updateParentVariantsInventoryInStockQuantity from "/imports/plugins/core/inventory/server/no-meteor/utils/updateParentVariantsInventoryInStockQuantity";
 import orderCreditMethod from "../util/orderCreditMethod";
+import appEvents from "/imports/node-app/core/util/appEvents";
+
 
 /**
  * @name orders/cancelOrder
@@ -29,78 +31,6 @@ export default function cancelOrder(order, returnToStock) {
   // Unless order is entirely contained in a single shop? Do we need a switch on marketplace owner dashboard?
   if (!Reaction.hasPermission("orders")) {
     throw new ReactionError("access-denied", "Access Denied");
-  }
-
-  // Inventory is removed from stock only once an order has been approved
-  // This is indicated by payment.status being anything other than `created`
-  // We need to check to make sure the inventory has been removed before we return it to stock
-  const orderIsApproved = order.shipping.find((group) => group.payment.status !== "created");
-
-  // If order is approved, the inventory has been taken away from both `inventoryQuantity` and `inventoryAvailableToSell`
-  if (returnToStock && orderIsApproved) {
-    // Run this Product update inline instead of using ordersInventoryAdjust because the collection hooks fail
-    // in some instances which causes the order not to cancel
-    const orderItems = order.shipping.reduce((list, group) => [...list, ...group.items], []);
-    orderItems.forEach(async (item) => {
-      if (Reaction.hasPermission("orders", Reaction.getUserId(), item.shopId)) {
-        Products.update(
-          {
-            _id: item.variantId,
-            shopId: item.shopId
-          },
-          {
-            $inc: {
-              inventoryAvailableToSell: +item.quantity,
-              inventoryQuantity: +item.quantity
-            }
-          },
-          {
-            bypassCollection2: true,
-            publish: true
-          }
-        );
-
-        // Update `inventoryAvailableToSell` on all parents of this variant / option
-        Promise.await(updateParentVariantsInventoryAvailableToSellQuantity(item, rawCollections));
-        // Update `inventoryQuantity` on all parents of this variant / option
-        Promise.await(updateParentVariantsInventoryInStockQuantity(item, rawCollections));
-
-        // Publish inventory updates to the Catalog
-        Promise.await(updateCatalogProductInventoryStatus(item.productId, rawCollections));
-      }
-    });
-  }
-
-  // If order is not approved, the inventory hasn't been taken away from `inventoryQuantity`, but has been taken away from `inventoryAvailableToSell`
-  if (!orderIsApproved) {
-    // Run this Product update inline instead of using ordersInventoryAdjust because the collection hooks fail
-    // in some instances which causes the order not to cancel
-    const orderItems = order.shipping.reduce((list, group) => [...list, ...group.items], []);
-    orderItems.forEach(async (item) => {
-      if (Reaction.hasPermission("orders", Reaction.getUserId(), item.shopId)) {
-        Products.update(
-          {
-            _id: item.variantId,
-            shopId: item.shopId
-          },
-          {
-            $inc: {
-              inventoryAvailableToSell: +item.quantity
-            }
-          },
-          {
-            bypassCollection2: true,
-            publish: true
-          }
-        );
-
-        // Update `inventoryAvailableToSell` on all parents of this variant / option
-        Promise.await(updateParentVariantsInventoryAvailableToSellQuantity(item, rawCollections));
-
-        // Publish inventory to catalog
-        Promise.await(updateCatalogProductInventoryStatus(item.productId, rawCollections));
-      }
-    });
   }
 
   const { _id: paymentId, invoice, paymentPluginName } = orderCreditMethod(order);
@@ -127,6 +57,8 @@ export default function cancelOrder(order, returnToStock) {
 
   // update item workflow
   Meteor.call("workflow/pushItemWorkflow", "coreOrderItemWorkflow/canceled", order, itemIds);
+
+  Promise.await(appEvents.emit("afterCancelOrder", order, returnToStock));
 
   return Orders.update(
     {
